+++ R/plot.check_normality.R
@@ -4,6 +4,7 @@
 #'
 #' @param type Character vector, indicating the type of plot.
 #' @param size_line Size of line geoms.
+#' @param alpha ALpha level of the confidence bands.
 #' @inheritParams data_plot
 #' @inheritParams plot.see_bayesfactor_parameters
 #'
@@ -17,7 +18,7 @@
 #' @importFrom bayestestR estimate_density
 #' @importFrom stats residuals sd dnorm rstudent ppoints pnorm fitted
 #' @export
-plot.see_check_normality <- function(x, type = c("density", "qq", "pp"), data = NULL, size_line = .8, size_point = 2, ...) {
+plot.see_check_normality <- function(x, type = c("density", "qq", "pp"), data = NULL, size_line = .8, size_point = 2, alpha = .2, ...) {
   type <- match.arg(type)
 
   if (is.null(data)) {
@@ -28,7 +29,7 @@ plot.see_check_normality <- function(x, type = c("density", "qq", "pp"), data = 
 
   # check type
   if (!is.null(attributes(x)$effects) && attributes(x)$effects == "random") {
-    .plot_diag_reqq(attributes(x)$re_qq, size_point = size_point, size_line = size_line)
+    .plot_diag_reqq(attributes(x)$re_qq, size_point = size_point, size_line = size_line, alpha_level = alpha)
   } else {
     if (type == "qq") {
       if (inherits(model, c("lme", "lmerMod", "merMod", "glmmTMB"))) {
@@ -38,7 +39,7 @@ plot.see_check_normality <- function(x, type = c("density", "qq", "pp"), data = 
       }
 
       dat <- stats::na.omit(data.frame(y = res_))
-      .plot_diag_qq(dat, size_point = size_point, size_line = size_line)
+      .plot_diag_qq(dat, size_point = size_point, size_line = size_line, alpha_level = alpha)
     } else if (type == "density") {
       r <- stats::residuals(model)
       dat <- as.data.frame(bayestestR::estimate_density(r))
@@ -51,7 +52,7 @@ plot.see_check_normality <- function(x, type = c("density", "qq", "pp"), data = 
     } else if (type == "pp") {
       x <- sort(stats::residuals(model), na.last = NA)
       dat <- data.frame(res = x)
-      .plot_diag_pp(dat, size_point = size_point, size_line = size_line)
+      .plot_diag_pp(dat, size_point = size_point, size_line = size_line, alpha_level = alpha)
     }
   }
 }